# openapi.yaml
openapi: 3.0.1
info:
  title: TODO Plugin
  description: A plugin that allows the user to create and manage a TODO list using ChatGPT. If you do not know the user's username, ask them first before making queries to the plugin. Otherwise, use the username "global".
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /qrCode/{username}:
    post:
      operationId: qrCode
      summary: Generate a QR code
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The username for which to generate a QR code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: The URL to generate a QR code from.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getQRResponse'
...
components:
  schemas:
    getQRResponse:
      type: object
      properties:
        qrCode:
          type: array
          items:
            type: string
          description: The string of qr.
...